feat(infinite queries): prefetch infiniteQueries with react-query server side
