+++ app/(admin)/dashboard/page.tsx
@@ -1,24 +1,49 @@
 // REACT
 import { Suspense } from "react";
+// DATA
+import { getCollectionData } from '@/data/firestore';
+// PACKAGES
+import { QueryClient, HydrationBoundary, dehydrate } from "@tanstack/react-query";
 // COMPONENTS
-import Items from "@/components/items/items";
 import { StatusFilters } from "@/components/status-filters";
 import { LayoutFilters } from "@/components/layout-filters";
+import { ItemsLayout } from "@/components/items/items-layout";
+// UTILS
+import { BASE_LIMIT } from "@/utils/constants";
 
 type SeaerchParams = {
   [key: string]: string | undefined;
 }
 
 export default async function DashboardPage({ searchParams }: { searchParams: SeaerchParams }) {
+  // SERVER LAND
+  const { status, limit, sort, display } = searchParams;
+  const queryClient = new QueryClient();
+  await queryClient.prefetchInfiniteQuery({
+    queryKey: ['queries', { status, limit, sort }],
+    initialPageParam: { 
+      status: status ?? 'ALL', 
+      limit: limit ?? BASE_LIMIT, 
+      sort: sort ?? 'desc', 
+      startAfter: null 
+    },
+    staleTime: 60 * 1000,
+    queryFn: async ({ pageParam }: { pageParam: any}) => {
+      return await getCollectionData('queries', pageParam);
+    },
+  })
+
   return (
     <>
       <h1>Dashboard</h1>
       <section className="flex flex-row justify-between items-stretch flex-wrap">
-        <StatusFilters status={searchParams.status || 'ALL'} />
-        <LayoutFilters layoutView={searchParams.display || 'GRID'} />
+        <StatusFilters status={status || 'ALL'} />
+        <LayoutFilters layoutView={display || 'GRID'} />
       </section>
       <Suspense fallback={(<h1>Loading ...</h1>)}>
-        <Items status={searchParams.status || 'ALL'} />
+        <HydrationBoundary state={dehydrate(queryClient)}>
+          <ItemsLayout />
+        </HydrationBoundary>
       </Suspense>
     </>
   );
